Separate queue management
